Added AMO Hardhat tasks
